mechanism_snabb: IPv4 support
This changes _choose_port behaviour. It first generates a list of all
ports that contain the requested zone with the given IP network version
(i.e. IPv4 or IPv6). Then this list goes is used to select the port
according to its load (gbps) availability.

The _choose_port tests are broken now. Will be fixed in the next patch.

Signed-off-by: Nikolay Nikolaev <n.nikolaev@virtualopensystems.com>