mech snabb: Track available IPv4 addresses more exactly
Fixes: https://github.com/virtualopensystems/neutron/issues/2

Currently we waste some IP addresses. This is not okay for IPv4 because so few addresses are available.

The issue is that we assign globally-unique "template addresses" (0.0.0.1, 0.0.0.2, 0.0.0.3, etc) and then move these into real subnets. However, it is overkill for these addresses to be globally unique. They only need to be unique within the real subnet that is eventually chosen.

For example, suppose we have three subnets (10.0.1.0/28, 10.0.2.0/28, 10.0.3.0/28) and we create one port in each one. Today the addresses we would assign are 10.0.1.1, 10.0.2.2, and 10.0.3.3. Instead we want to allocate 10.0.1.1, 10.0.2.1, and 10.0.3.1 i.e. the first available addresses for each subnet individually.

Signed-off-by: Nikolay Nikolaev <n.nikolaev@virtualopensystems.com>